[OPIK-4173] [FE] Move thread stats fetching into useProjectTabs hook
