+++ apps/opik-frontend/src/components/pages/TracesPage/TracesPage.tsx
@@ -1,10 +1,7 @@
-import { useMemo } from "react";
 import { StringParam, useQueryParam } from "use-query-params";
 import { useProjectIdFromURL } from "@/hooks/useProjectIdFromURL";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import useProjectById from "@/api/projects/useProjectById";
-import useThreadsStatistic from "@/api/traces/useThreadsStatistic";
-import { useMetricDateRangeWithQueryAndStorage } from "@/components/pages-shared/traces/MetricDateRangeSelect";
 import PageBodyScrollContainer from "@/components/layout/PageBodyScrollContainer/PageBodyScrollContainer";
 import PageBodyStickyContainer from "@/components/layout/PageBodyStickyContainer/PageBodyStickyContainer";
 import LogsTab from "@/components/pages/TracesPage/LogsTab/LogsTab";
@@ -25,7 +22,6 @@ import ViewSelector, {
 } from "@/components/pages-shared/dashboards/ViewSelector/ViewSelector";
 import useProjectTabs from "@/components/pages/TracesPage/useProjectTabs";
 import { PROJECT_TAB } from "@/constants/traces";
-import { STATISTIC_AGGREGATION_TYPE } from "@/types/shared";
 
 const TracesPage = () => {
   const projectId = useProjectIdFromURL();
@@ -46,35 +42,6 @@ const TracesPage = () => {
 
   const projectName = project?.name || projectId;
 
-  const { intervalStart, intervalEnd } =
-    useMetricDateRangeWithQueryAndStorage();
-
-  const { data: threadsStats } = useThreadsStatistic(
-    {
-      projectId,
-      fromTime: intervalStart,
-      toTime: intervalEnd,
-    },
-    {
-      enabled: !!projectId,
-      refetchOnMount: false,
-    },
-  );
-
-  const threadCount = useMemo(() => {
-    if (!threadsStats) return undefined;
-
-    const threadCountStat = threadsStats.stats?.find(
-      (stat) =>
-        stat.name === "thread_count" &&
-        stat.type === STATISTIC_AGGREGATION_TYPE.COUNT,
-    );
-
-    return threadCountStat?.type === STATISTIC_AGGREGATION_TYPE.COUNT
-      ? threadCountStat.value
-      : 0;
-  }, [threadsStats]);
-
   const {
     activeTab,
     logsType,
@@ -83,7 +50,6 @@ const TracesPage = () => {
     handleTabChange,
   } = useProjectTabs({
     projectId,
-    threadCount,
   });
 
   const [view = VIEW_TYPE.DETAILS, setView] = useQueryParam(
